[INJIMOB-3598] refactor: separate data classes to different files
