+++ Sources/VCIClient/authorizationCodeFlow/interactiveAuthorization/AuthorizationRequestData.swift
@@ -1,0 +1,3 @@
+import Foundation
+
+protocol AuthorizationRequestData {}
