+++ Sources/VCIClient/authorizationCodeFlow/interactiveAuthorization/PresentationAuthorizationRequestData.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+struct PresentationAuthorizationRequestData: AuthorizationRequestData {
+    let ovpRequest: [String: Any]
+    let authSession: String?
+    let iar: String
+}
